requirements.txt
Flask
gunicorn
PyPDF2
python-dotenv
openai